requirements.txt
streamlit==1.30.0
pdfplumber==0.10.0
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.0.0
